refactor(commands): hoist stub helpers into a trait
