+++ src/Commands/MakeControllerCommand.php
@@ -5,9 +5,12 @@ namespace PlinCode\LaravelCleanArchitecture\Commands;
 use Illuminate\Console\Command;
 use Illuminate\Filesystem\Filesystem;
 use Illuminate\Support\Str;
+use PlinCode\LaravelCleanArchitecture\Concerns\RendersStubs;
 
 class MakeControllerCommand extends Command
 {
+    use RendersStubs;
+
     protected $signature = 'clean-arch:make-controller {name : The name of the controller}
                           {--api : Generate API controller}
                           {--web : Generate Web controller}
@@ -90,16 +93,5 @@ class MakeControllerCommand extends Command
         ];
 
         return str_replace(array_keys($replacements), array_values($replacements), $content);
-    }
-
-    protected function getStub(string $stub): string
-    {
-        $stubPath = __DIR__ . "/../../stubs/{$stub}.stub";
-
-        if (! $this->files->exists($stubPath)) {
-            throw new \Exception("Stub file not found: {$stubPath}");
-        }
-
-        return $this->files->get($stubPath);
     }
 }
