+++ src/Commands/MakeJobCommand.php
@@ -4,10 +4,12 @@ namespace PlinCode\LaravelCleanArchitecture\Commands;
 
 use Illuminate\Console\Command;
 use Illuminate\Filesystem\Filesystem;
-use Illuminate\Support\Str;
+use PlinCode\LaravelCleanArchitecture\Concerns\RendersStubs;
 
 class MakeJobCommand extends Command
 {
+    use RendersStubs;
+
     protected $signature = 'clean-arch:make-job {name : The name of the job}
                           {--force : Overwrite existing files}';
 
@@ -35,7 +37,7 @@ class MakeJobCommand extends Command
     protected function createJob(string $name): void
     {
         $stub    = $this->getStub('job');
-        $content = $this->replacePlaceholders($stub, $name);
+        $content = $this->replaceDomainPlaceholders($stub, $name);
 
         $jobPath = app_path('Application/Jobs');
 
@@ -45,28 +47,5 @@ class MakeJobCommand extends Command
 
         $this->files->put("{$jobPath}/{$name}Job.php", $content);
         $this->info("Created: Application/Jobs/{$name}Job.php");
-    }
-
-    protected function replacePlaceholders(string $content, string $name): string
-    {
-        $pluralName     = Str::plural($name);
-        $domainVariable = Str::camel($name);
-
-        return str_replace(
-            ['{{DomainName}}', '{{PluralDomainName}}', '{{domainVariable}}'],
-            [$name, $pluralName, $domainVariable],
-            $content
-        );
-    }
-
-    protected function getStub(string $stub): string
-    {
-        $stubPath = __DIR__ . "/../../stubs/{$stub}.stub";
-
-        if (! $this->files->exists($stubPath)) {
-            throw new \Exception("Stub file not found: {$stubPath}");
-        }
-
-        return $this->files->get($stubPath);
     }
 }
