+++ src/Commands/MakeListenerCommand.php
@@ -4,10 +4,12 @@ namespace PlinCode\LaravelCleanArchitecture\Commands;
 
 use Illuminate\Console\Command;
 use Illuminate\Filesystem\Filesystem;
-use Illuminate\Support\Str;
+use PlinCode\LaravelCleanArchitecture\Concerns\RendersStubs;
 
 class MakeListenerCommand extends Command
 {
+    use RendersStubs;
+
     protected $signature = 'clean-arch:make-listener {name : The name of the listener}
                           {--force : Overwrite existing files}';
 
@@ -35,7 +37,7 @@ class MakeListenerCommand extends Command
     protected function createListener(string $name): void
     {
         $stub    = $this->getStub('listener');
-        $content = $this->replacePlaceholders($stub, $name);
+        $content = $this->replaceDomainPlaceholders($stub, $name);
 
         $listenerPath = app_path('Application/Listeners');
 
@@ -45,28 +47,5 @@ class MakeListenerCommand extends Command
 
         $this->files->put("{$listenerPath}/{$name}Listener.php", $content);
         $this->info("Created: Application/Listeners/{$name}Listener.php");
-    }
-
-    protected function replacePlaceholders(string $content, string $name): string
-    {
-        $pluralName     = Str::plural($name);
-        $domainVariable = Str::camel($name);
-
-        return str_replace(
-            ['{{DomainName}}', '{{PluralDomainName}}', '{{domainVariable}}'],
-            [$name, $pluralName, $domainVariable],
-            $content
-        );
-    }
-
-    protected function getStub(string $stub): string
-    {
-        $stubPath = __DIR__ . "/../../stubs/{$stub}.stub";
-
-        if (! $this->files->exists($stubPath)) {
-            throw new \Exception("Stub file not found: {$stubPath}");
-        }
-
-        return $this->files->get($stubPath);
     }
 }
