+++ src/Commands/MakeNotificationCommand.php
@@ -4,10 +4,12 @@ namespace PlinCode\LaravelCleanArchitecture\Commands;
 
 use Illuminate\Console\Command;
 use Illuminate\Filesystem\Filesystem;
-use Illuminate\Support\Str;
+use PlinCode\LaravelCleanArchitecture\Concerns\RendersStubs;
 
 class MakeNotificationCommand extends Command
 {
+    use RendersStubs;
+
     protected $signature = 'clean-arch:make-notification {name : The name of the notification}
                           {--force : Overwrite existing files}';
 
@@ -35,7 +37,7 @@ class MakeNotificationCommand extends Command
     protected function createNotification(string $name): void
     {
         $stub    = $this->getStub('notification');
-        $content = $this->replacePlaceholders($stub, $name);
+        $content = $this->replaceDomainPlaceholders($stub, $name);
 
         $notificationPath = app_path('Infrastructure/Notifications');
 
@@ -45,28 +47,5 @@ class MakeNotificationCommand extends Command
 
         $this->files->put("{$notificationPath}/{$name}Notification.php", $content);
         $this->info("Created: Infrastructure/Notifications/{$name}Notification.php");
-    }
-
-    protected function replacePlaceholders(string $content, string $name): string
-    {
-        $pluralName     = Str::plural($name);
-        $domainVariable = Str::camel($name);
-
-        return str_replace(
-            ['{{DomainName}}', '{{PluralDomainName}}', '{{domainVariable}}'],
-            [$name, $pluralName, $domainVariable],
-            $content
-        );
-    }
-
-    protected function getStub(string $stub): string
-    {
-        $stubPath = __DIR__ . "/../../stubs/{$stub}.stub";
-
-        if (! $this->files->exists($stubPath)) {
-            throw new \Exception("Stub file not found: {$stubPath}");
-        }
-
-        return $this->files->get($stubPath);
     }
 }
