+++ src/Commands/MakeObserverCommand.php
@@ -5,9 +5,12 @@ namespace PlinCode\LaravelCleanArchitecture\Commands;
 use Illuminate\Console\Command;
 use Illuminate\Filesystem\Filesystem;
 use Illuminate\Support\Str;
+use PlinCode\LaravelCleanArchitecture\Concerns\RendersStubs;
 
 class MakeObserverCommand extends Command
 {
+    use RendersStubs;
+
     protected $signature = 'clean-arch:make-observer {name : The name of the observer}
                           {domain : The domain name}
                           {--force : Overwrite existing files}';
@@ -60,16 +63,5 @@ class MakeObserverCommand extends Command
             [$domain, $pluralDomain, $domainVariable],
             $content
         );
-    }
-
-    protected function getStub(string $stub): string
-    {
-        $stubPath = __DIR__ . "/../../stubs/{$stub}.stub";
-
-        if (! $this->files->exists($stubPath)) {
-            throw new \Exception("Stub file not found: {$stubPath}");
-        }
-
-        return $this->files->get($stubPath);
     }
 }
