+++ src/Commands/MakeServiceCommand.php
@@ -67,15 +67,4 @@ class MakeServiceCommand extends Command
 
         return str_replace(array_keys($replacements), array_values($replacements), $content);
     }
-
-    protected function getStub(string $stub): string
-    {
-        $stubPath = __DIR__ . "/../../stubs/{$stub}.stub";
-
-        if (! $this->files->exists($stubPath)) {
-            throw new \Exception("Stub file not found: {$stubPath}");
-        }
-
-        return $this->files->get($stubPath);
-    }
 }
